Transformation Structure
I've extended the ITransform interface to support binary input via the
BinaryReader class. I think most instances that only support text should
be able to pass the underlying stream to the overload that takes a
TextReader instance. We may want to implement a base class to remove
some of the boilerplate. I have also removed some of the overloads from
TransformationExtensions that return Task instances.